src/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
argon2-cffi==21.3.0
    # via
    #   -r requirements.in
    #   flask-argon2
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
async-timeout==4.0.2
    # via redis
boto3==1.23.5
    # via -r requirements.in
botocore==1.26.5
    # via
    #   boto3
    #   s3transfer
cffi==1.15.0
    # via argon2-cffi-bindings
click==8.1.2
    # via
    #   flask
    #   rq
deprecated==1.2.13
    # via redis
flask==2.1.1
    # via
    #   -r requirements.in
    #   flask-and-redis
    #   flask-argon2
    #   flask-login
    #   flask-sqlalchemy
flask-and-redis==1.0.0
    # via -r requirements.in
flask-argon2==0.2.0.0
    # via -r requirements.in
flask-login==0.6.0
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via -r requirements.in
greenlet==1.1.2
    # via sqlalchemy
importlib-metadata==4.11.3
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.1
    # via flask
jmespath==1.0.0
    # via
    #   boto3
    #   botocore
markupsafe==2.1.1
    # via jinja2
marshmallow==3.15.0
    # via -r requirements.in
packaging==21.3
    # via
    #   marshmallow
    #   redis
phonenumbers==8.12.50
    # via -r requirements.in
psycopg2==2.9.3
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyparsing==3.0.8
    # via packaging
python-dateutil==2.8.2
    # via botocore
python-slugify==6.1.2
    # via -r requirements.in
redis==4.2.2
    # via
    #   -r requirements.in
    #   flask-and-redis
    #   rq
rq==1.10.1
    # via -r requirements.in
s3transfer==0.5.2
    # via boto3
shortuuid==1.0.8
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlalchemy==1.4.35
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
text-unidecode==1.3
    # via python-slugify
urllib3==1.26.9
    # via botocore
werkzeug==2.1.1
    # via
    #   flask
    #   flask-login
wrapt==1.14.0
    # via deprecated
zipp==3.8.0
    # via importlib-metadata
